Split iOS SQLite review tests
